Factor RemoteParam's keys into a ChannelPubKeys
We already have a type (ChannelPubKeys) for holding the public keys for a
channel end. The LocalParams type uses this to store the keys of the local end
of a channel, but RemoteParams instead inlines all the keys for the remote end
directly in the RemoteParams type. For consistency and simplicity, move these
keys into a ChannelPubKeys stored in the RemoteParams.