Replace GetSignatureFor with SignWithFundingPrivKey
We only use this function to sign with the funding key. Making that
explicit means we can avoid passing in the key (since the key repo
already stores the funding key) and avoid downcasting the key using
RawPubKey.

Similar functions could be added to sign with other keys if need be.