Remove IKeysRepository
IKeysRepository duplicates the functionality of the ChannelPrivKeys
type. As I understand it, this type was originally supposed to act as a
repository of all the user's keys, but since it is constructed with a
channel index it only actually stores the keys for a specific channel.
Since this is also what ChannelPrivKeys does, there's no need to have
both.

The other thing that IKeysRepository does is store the master private
key for the node. This key is used to derive the key for the TCP crypto
layer (ie. it determines the node id) though we also use it to derive
channel-specific keys (though it's not used for that in IKeysRepository
after the IKeysRepository is constructed). As such, I've split this into
a separate NodeMasterPrivKey type. I've also added a separate NodeSecret
type (the TCP crypto key, derived from NodeMasterPrivKey) since the node
master private key and node secrets were being conflated (ie. the term
"node secret" was being used throughout the code to refer to both keys
interchangeably).