Turn key derivation functions into methods
Each key type now has a separate method for deriving it. This is to
prevent keys from being misused as the wrong kind of key.

In the lightning network every commitment tx has a per commitment point
which determines how the commitment's keys are derived from the
basepoints. Since the PerCommitmentPoint can be seen as a kind of
"master key" for a commitment, the functions methods have been turned
into methods on that type. This makes the code simpler and easier to
understand.