Remove pre-channel states
These states don't have commitments, only exist before the existence of
a channel is confirmed, don't need to be saved in a wallet, and can only
handle one specific ChannelCommand each.

As such, it doesn't make much sense to have them be part of
ChannelState. They can instead be seperate types which exist prior to the
creation of a channel and which have specialized methods for performing
the single operation which they can handle.