Move RemoteNextCommitInfo into the channel state
This field lived in Commitments but we don't always have this
information available when we have a Commitments object, so it was being
filled with dummy data.

We only obtain the remote's next per commitment point when we receive
funding_locked. As such, this field should only be present in states
that are subsequent to funding_locked.